Created func to retreive data
